# Improve performance by using native `find` and `egrep`