As I was profiling importing files on a fairly large (Brigade) codebase,
I found a couple of interesting things:

  a) The `Dir.glob` call used to find all .js* files was taking about
     0.2 seconds.
  b) The loop to find exactly those files to import took about 0.4
     seconds too.
  c) The check to see if anything had changed (file added/removed) in a
     directory took about 0.1 seconds.
  d) In total, an import would take between 0.6 and 0.7 seconds,
     independent of whether the cache was used or not.

My previous attempt at improving performance through a cache layer
didn't make too much of a difference, because I was optimizing (a). I
should have tried to optimize (b) instead if I had been smart. This
second attempt at improving performance, addresses both (a) and (b) by
delegating the `Dir.glob` to a external `find` command, and the
filtering loop to an external `egrep` command. This takes importing
times down to under 0.2 seconds on my laptop, so about three times
faster than before, and without any complicated cache.